refactor DirectoryListItem
